relocate io mp -> utilities
